+++ src/js/routes/Vote.jsx
@@ -14,7 +14,8 @@ import BallotStore from '../stores/BallotStore';
 import BallotTitleHeader from './Ballot/BallotTitleHeader';
 import BrowserPushMessage from '../components/Widgets/BrowserPushMessage';
 import cookies from '../utils/cookies';
-import { cordovaVoteMiniHeader, cordovaScrollablePaneTopPadding } from '../utils/cordovaOffsets';
+import { cordovaVoteMiniHeader } from '../utils/cordovaOffsets';
+import cordovaScrollablePaneTopPadding from '../utils/cordovaScrollablePaneTopPadding';
 import { historyPush, isCordova, isWebApp } from '../utils/cordovaUtils';
 import ElectionActions from '../actions/ElectionActions';
 import FindPollingLocation from '../components/Vote/FindPollingLocation';
