+++ src/js/startCordova.js
@@ -1,5 +1,7 @@
 import TwitterSignIn from './components/Twitter/TwitterSignIn';
-import { isAppleSilicon, isCordova, isIOS, prepareForCordovaKeyboard, restoreStylesAfterCordovaKeyboard } from './utils/cordovaUtils';
+import { dumpScreenAndDeviceFields } from './utils/appleSiliconUtils';
+import { getProcessorArchitecture, isIOSAppOnMac, isCordova, isIOS,
+  prepareForCordovaKeyboard, restoreStylesAfterCordovaKeyboard } from './utils/cordovaUtils';
 import VoterActions from './actions/VoterActions';
 
 function localPrepareForCordovaKeyboard () {
@@ -25,6 +27,19 @@ export function initializationForCordova () { // eslint-disable-line
   const { cordova: { InAppBrowser } } = window;
   window.open = InAppBrowser.open;
 
+  getProcessorArchitecture();
+
+  const { cordova: { getAppVersion: getVersionNumber } } = window;
+  getVersionNumber().then((version) => {
+    window.weVoteAppVersion = version;
+    return true;
+  });
+
+
+  if (isIOSAppOnMac()) {
+    dumpScreenAndDeviceFields();
+  }
+
   // Special keyboard handling for iOS
   if (isIOS()) {
     // Unfortunately this event only works on iOS, but fortunately it is most needed on iOS
@@ -32,7 +47,7 @@ export function initializationForCordova () { // eslint-disable-line
     window.addEventListener('keyboardDidHide', localRestoreStylesAfterCordovaKeyboard);
   }
 
-  if (isCordova() && !isAppleSilicon()) {
+  if (isCordova() && !isIOSAppOnMac()) {
     const { cordova: { plugins: { firebase: { messaging } } } } = window;
     // https://github.com/chemerisuk/cordova-plugin-firebase-messaging
     // For iOS, this can't be tested in a simulator.  Works fine in simulator on Android.
