+++ src/js/stores/OrganizationStore.js
@@ -503,7 +503,7 @@ class OrganizationStore extends ReduceStore {
           allCachedOrganizationsDict[organizationWeVoteId] = organization;
           // Now request fresh siteConfiguration data if the siteOwner was updated
           ({ hostname } = window.location);
-          hostname = hostname || 'silicon';
+          hostname = hostname || '';
           // console.log('OrganizationStore organizationPhotosSave hostname:', hostname, ', organizationWeVoteId:', organizationWeVoteId);
           AppActions.siteConfigurationRetrieve(hostname);
           return {
