+++ src/js/utils/appleSiliconUtils.js
@@ -1,4 +1,4 @@
-
+// Misc utilities to help with Apple Silicon debugging, where the Safari Web Inspector is not available.
 
 export function dumpCssFromId (id) {
   const el = document.getElementById(id);
@@ -14,4 +14,9 @@ export function dumpCssFromId (id) {
 export function dumpObjProps (name, obj) {
   // eslint-disable-next-line guard-for-in
   Object.keys(obj).forEach((key) => console.log(`Dump Object ${name} ${key}: ${obj[key]}`));
+}
+
+export function dumpScreenAndDeviceFields () {
+  dumpObjProps('window.screen', window.screen);
+  dumpObjProps('window.device', window.device);
 }
